+++ src/RideWindow/RideWindowFunctions.ts
@@ -1,5 +1,5 @@
-import { showWindowRide } from "./RideWindow";
+import {showWindowRide} from "./RideWindow";
+
 export function openRideWindow(item: Ride) {
-	let ride: Ride = item;
-	showWindowRide(ride);
+    showWindowRide(item);
 }