+++ CHANGELOG.md
@@ -48,3 +48,20 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 - Setting the colorscheme of a ride
 - Ridewindow following a train
 - Saving groups in file
+
+## [M2 ThrillTracker V0.0.4] - 2024-06-14
+### Added
+- Added the ability to rename a group
+- Added the ability to save a group to a file
+- Added custom error messages
+
+### Changed
+- Changed the way the list of rides is displayed
+
+### Removed
+- No Removed changes planned for M2
+
+### Planned features for beta release
+- final touches to the ride window
+- rework the windows when a ride/group is renamed/removed.
+